VNCTF2023/misc-lsstib/app/requirements.txt
Flask
pillow
numpy
pycrypto